+++ package.yaml
@@ -38,6 +38,8 @@ default-extensions:
   - UndecidableInstances
   - TypeApplications
   - FlexibleContexts
+  - LambdaCase
+  - TemplateHaskellQuotes
 
 when:
   - condition: impl(ghc < 8.6)
